Break up package tool monolith.
Break out logical modules and individual test functions.
